Check before replacing .vimrc

+++ setup.sh
@@ -1,14 +1,16 @@
 #!/bin/bash
-CYAN='\033[1;36m'
-PURPLE='\033[1;35m'
-WHITE='\033[0;37m'
-YELLOW='\033[1;33m'
 
-echo -e "${PURPLE}Installing .vimrc...${YELLOW}"
+echo -e "Installing .vimrc..."
 ln -s ~/.vim/.vimrc ~/.vimrc
 
 if [ "$?" -ne "0" ]; then
-  echo -e "${PURPLE}~/.vimrc already exists, removing..."
-  rm ~/.vimrc
-  ln -s ~/.vim/.vimrc ~/.vimrc
+  read -p "~/.vimrc already exists. Would you like to overwrite it? [y/n] " prompt
+  if [[ $prompt == "y" || $prompt == "Y" || $prompt == "yes" || $prompt == "Yes" ]]
+  then
+    rm ~/.vimrc;
+    ln -s ~/.vim/.vimrc ~/.vimrc;
+    echo -e "~/.vimrc replaced with ~/.vim/.vimrc";
+  else
+    exit;
+  fi
 fi
